+++ src/App.jsx
@@ -4,6 +4,7 @@ import { Routes, Route } from 'react-router-dom';
 import Menu from '../components/Menu'
 import './App.css'
 import Home from '../components/Home';
+import ArticleProfile from '../components/ArticleProfile'
 
 import { ArticlesProvider } from '../contexts/ArticlesContext';
 function App() {
@@ -17,7 +18,7 @@ function App() {
     <Routes>
       <Route path="/" element={<Home/>} />
       <Route path="/api/articles" element={<ArticleList/>} /> 
-      <Route path="/articles/:article_id" element={<UserProfile/>} />
+      <Route path="/api/articles/:article_id" element={<ArticleProfile/>} />
     </Routes>
     </ArticlesProvider>
